# Conflicts:
#	Front/PetPlace/app/src/main/java/com/example/petplace/di/NetworkModule.kt

+++ Back/petPlace/src/main/java/com/minjeok4go/petplace/chat/dto/ChatRoomDTO.java
@@ -9,9 +9,9 @@ import java.time.LocalDateTime;
 @Getter@Setter
 @AllArgsConstructor
 public class ChatRoomDTO {
-    private Integer chatRoomId;
-    private Integer userId1;
-    private Integer userId2;
+    private Long chatRoomId;
+    private Long userId1;
+    private Long userId2;
     private String lastMessage;
     private LocalDateTime lastMessageAt;
 }
